LPS-58615 change count to long instead of int

+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/internal/stats/DefaultStatsTranslator.java
@@ -57,7 +57,7 @@ public class DefaultStatsTranslator implements StatsTranslator {
 			ValueCount valueCount = (ValueCount)aggregationMap.get(
 				field + "_count");
 
-			statsResults.setCount((int)valueCount.getValue());
+			statsResults.setCount(valueCount.getValue());
 		}
 
 		if (stats.isMax()) {
